refactor(web): extract shared SelectorPicker component (#936)
Lift the correlation `SelectorPicker` (kind dropdown + per-kind autocompleted
inputs) out of the Correlations explorer into a reusable
`components/SelectorPicker.tsx` (with its `KIND_LABELS`/`NUTRIENTS` and a
co-located CSS, plus an exported `ALL_SELECTOR_KINDS`). The explorer now imports
it; behaviour is unchanged. This is the picker the article correlation-block
composer will reuse, so both surfaces build selectors the same way.
