+++ apps/web/src/components/SelectorPicker.css
@@ -1,0 +1,14 @@
+.selector-picker {
+  display: flex;
+  gap: 8px;
+  flex-wrap: wrap;
+  align-items: center;
+}
+
+.selector-picker select,
+.selector-picker input {
+  padding: 6px 8px;
+  border: 1px solid #d1d5db;
+  border-radius: 6px;
+  font-size: 0.9rem;
+}
